requirements.txt
Flask>=1.0
mysql-connector-python>=8.0.20
